+++ src/main/java/com/PorfolioArgPrograma/Porfolio/Repository/ProyectoRepository.java
@@ -10,6 +10,6 @@ import org.springframework.stereotype.Repository;
  */
 
 @Repository
-public interface ProyectoRepository extends JpaRepository<Proyecto, Integer> {
+public interface ProyectoRepository extends JpaRepository<Proyecto, Long> {
     
 }
