+++ src/main/java/com/PorfolioArgPrograma/Porfolio/Repository/SkillRepository.java
@@ -9,6 +9,6 @@ import org.springframework.stereotype.Repository;
  * @author Juan Pablo
  */
 @Repository
-public interface SkillRepository extends JpaRepository<Skill, Integer> {
+public interface SkillRepository extends JpaRepository<Skill, Long> {
     
 }
